criu: use proper format-specified to accommodate time_t 64-bit change
See also:
https://wiki.debian.org/ReleaseGoals/64bit-time

+++ criu/autofs.c
@@ -658,7 +658,7 @@ static int autofs_mnt_make_catatonic(const char *mnt_path, int mnt_fd)
 
 static int autofs_mnt_set_timeout(time_t timeout, const char *mnt_path, int mnt_fd)
 {
-	pr_info("%s: set timeout %ld for %s\n", __func__, timeout, mnt_path);
+	pr_info("%s: set timeout %" PRId64 " for %s\n", __func__, (int64_t)timeout, mnt_path);
 	return autofs_ioctl(mnt_path, mnt_fd, AUTOFS_IOC_SETTIMEOUT, &timeout);
 }
 
@@ -770,7 +770,7 @@ static int autofs_post_mount(const char *mnt_path, dev_t mnt_dev, time_t timeout
 	}
 
 	if (autofs_mnt_set_timeout(timeout, mnt_path, mnt_fd)) {
-		pr_err("Failed to set timeout %ld for %s\n", timeout, mnt_path);
+		pr_err("Failed to set timeout %" PRId64 " for %s\n", (int64_t)timeout, mnt_path);
 		return -1;
 	}
